AuthVista/backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
geoalchemy2==0.14.3
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-socketio==5.11.0
aiofiles==23.2.1
redis==5.0.1
celery==5.3.4

# YOLO Inference
ultralytics==8.1.11
opencv-python==4.9.0.80
torch==2.1.2
torchvision==0.16.2
pillow==10.2.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
